web-interface/requirements.txt
flask
pandas
psycopg2-binary
requests